refactor institution list
