+++ src/app/app.module.ts
@@ -42,8 +42,8 @@ import {MAT_DATE_LOCALE, MatNativeDateModule} from '@angular/material/core'
 import {MatTooltipModule} from '@angular/material/tooltip'
 import {ZsbReportComponent} from './zsb-events/zsb-report/zsb-report.component'
 import {ZsbInstitutionsComponent} from './zsb-institutions/zsb-institutions.component'
-import {ZsbInstitutionsListComponent} from './zsb-institutions/zsb-institutionen-list/zsb-institutions-list.component'
-import {ZsbInstitutionsDetailComponent} from './zsb-institutions/zsb-institutionen-detail/zsb-institutions-detail.component'
+import {ZsbInstitutionsListComponent} from './zsb-institutions/zsb-institutions-list/zsb-institutions-list.component'
+import {ZsbInstitutionsDetailComponent} from './zsb-institutions/zsb-institutions-detail/zsb-institutions-detail.component'
 import {ZsbLetterComponent} from './zsb-communication/zsb-letter/zsb-letter.component'
 import {ZsbEmailComponent} from './zsb-communication/zsb-email/zsb-email.component'
