+++ src/app/shared/database.service.ts
@@ -7,7 +7,7 @@ import {Observable, throwError} from 'rxjs'
 import {SchoolType} from '../zsb-school/schoolType'
 import {City} from '../zsb-address/city'
 import {AmountStudents} from '../zsb-school/amount-students'
-import {Contact, ContactSalutation, ContactFunction} from '../zsb-contact/contact'
+import {Contact, ContactFunction, ContactSalutation} from '../zsb-contact/contact'
 import {environment} from '../../environments/environment'
 import {Event} from '../zsb-events/event'
 import {Category} from '../zsb-events/category'
